Allow keyboard user input for prompts
A charachter for testing is placed on the tile 7,10 of the first map.
RS_InputDialog is used to capture the user input and it's fairly customizable.
If \V[x] is passed as the prompt, the agent plugin will take the content of a variable with the ID x as the prompt.